# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Telecom Management - API Documentation",
        "version": "1.0"
    },
    "paths": {
...
        "/api/cucm/line": {
            "put": {
                "tags": [
                    "Management - CUCM - Line - Provisioning"
                ],
                "summary": "Update Line in CUCM",
                "description": "Update Line",
                "operationId": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "line",
                        "in": "formData",
                        "description": "This requires the correct values to be passed.",
                        "required": true,
                        "type": "array",
                        "items": {
                            "description": "Phone",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                }
            },
            "post": {
                "tags": [
                    "Management - CUCM - Phone Provisioning"
                ],
                "summary": "Create New Line in CUCM",
                "description": "",
                "operationId": "createSite",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "sitecode",
                        "in": "formData",
                        "description": "Sitecode",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "device",
                        "in": "formData",
                        "description": "Device Type - Example: 7945, 8841, IP Communicator",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "Phone Name - Example 0004DEADBEEF",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "firstname",
                        "in": "formData",
                        "description": "First Name - John",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "lastname",
                        "in": "formData",
                        "description": "Last Name - Doe",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "username",
                        "in": "formData",
                        "description": "User Name - John.Doe, CallManager.Unassign",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "dn",
                        "in": "formData",
                        "description": "Directory Number - Example: 4025551234",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "extlength",
                        "in": "formData",
                        "description": "Internal Extension Length - 4 digit is standard - Used for Internal Short Dialing",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "4",
                            "5",
                            "10"
                        ]
                    },
                    {
                        "name": "language",
                        "in": "formData",
                        "description": "Language",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "English",
                            "French"
                        ]
                    },
                    {
                        "name": "voicemail",
                        "in": "formData",
                        "description": "Does user require a Voicemail Box with this DN?",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "true",
                            "false"
                        ]
                    },
                    {
                        "name": "notes",
                        "in": "formData",
                        "description": "Notes",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                }
            }
        },
...
        "/api/cucm/phone/{name}": {
            "get": {
                "tags": [
                    "Management - CUCM - Phone Provisioning"
                ],
                "summary": "Get Phone Details by Name",
                "description": "",
                "operationId": "getPhone",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Name of Phone - Example SEP0004DEADBEEF",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Management - CUCM - Phone Provisioning"
                ],
                "summary": "Remove Phone by Name",
                "description": "",
                "operationId": "deletePhone",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Name of Phone to Delete - Example SEP0004DEADBEEF",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
        "/api/cucm/phone": {
            "put": {
                "tags": [
                    "Management - CUCM - Phone Provisioning"
                ],
                "summary": "Update Phone in CUCM",
                "description": "Update Phone",
                "operationId": "insert phones",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "This requires the correct values to be passed.",
                        "required": true,
                        "type": "array",
                        "items": {
                            "description": "Phone",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                }
            },
            "post": {
                "tags": [
                    "Management - CUCM - Phone Provisioning"
                ],
                "summary": "Create New Phone in CUCM",
                "description": "",
                "operationId": "createSite",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "sitecode",
                        "in": "formData",
                        "description": "Sitecode",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "device",
                        "in": "formData",
                        "description": "Device Type - Example: 7945, 8841, IP Communicator",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "Phone Name - Example 0004DEADBEEF",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "firstname",
                        "in": "formData",
                        "description": "First Name - John",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "lastname",
                        "in": "formData",
                        "description": "Last Name - Doe",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "username",
                        "in": "formData",
                        "description": "User Name - John.Doe, CallManager.Unassign",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "dn",
                        "in": "formData",
                        "description": "Directory Number - Example: 4025551234",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "extlength",
                        "in": "formData",
                        "description": "Internal Extension Length - 4 digit is standard - Used for Internal Short Dialing",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "4",
                            "5",
                            "10"
                        ]
                    },
                    {
                        "name": "language",
                        "in": "formData",
                        "description": "Language",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "English",
                            "French"
                        ]
                    },
                    {
                        "name": "voicemail",
                        "in": "formData",
                        "description": "Does user require a Voicemail Box with this DN?",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "true",
                            "false"
                        ]
                    },
                    {
                        "name": "notes",
                        "in": "formData",
                        "description": "Notes",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                }
            }
        },
...
        "/api/cucm/upload/phones": {
            "post": {
                "tags": [
                    "Management - CUCM - Phone Provisioning"
                ],
                "summary": "Upload Phone Planning Document in CUCM",
                "description": "Upload .csv file of phones",
                "operationId": "createSite",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "phones",
                        "in": "formData",
                        "description": "Sitecode",
                        "required": true,
                        "type": "file"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                }
            }
        },
...
        "/api/cucm/site/migration/run": {
            "post": {
                "tags": [
                    "Management - CUCM - Site Migration"
                ],
                "summary": "Run Migration Plan",
                "description": "Run Migration",
                "operationId": "createSite",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "type",
                        "in": "formData",
                        "description": "Add, Update, Delete",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "migration",
                        "in": "formData",
                        "description": "Array of Migration Ojbects",
                        "required": true,
                        "type": "array",
                        "items": {
                            "description": "cucm objects",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                }
            }
        },
...
    },
...
}